Split native Zenmap profile editing helpers
